[mtp-configfs] Allow specifying the USBMODE via multi-instance
This might abuse a bit the multi-instance functionality, but allows us
to avoid changing the configuration file in the first place, that can stay
owned by root and unchangeable (also, since it was actually fully sourced
inside mtp-configfs, it was a real security issue that could potentially
lead to privilege escalation).

Extra care should be taken when switching (old instance must be disabled,
new one should be enabled and started).

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>